lib/efi: Add efi_main and support
We can launch the unit tests from UEFI by embedding them in EFI
apps. gnu-efi apps start in efi_main(). Add that and all its
supporting functions that allow us to call a unit test's setup(),
main(), and exit() functions.

When EFI_DEBUG is not defined we redefine ASSERT in a way that
eats its input. This ensures that variables which are only used
by ASSERT do not generate unused warnings. The warnings were
reported by Nikos Nikoleris.

Additionally reported by Nikos Nikoleris was that the stack UEFI
provides is too small. The efi_allocate_stack() function and the
introduction of primary_entry(), which is where the stack may be
switched, are based on a patch from Nikos.

Signed-off-by: Andrew Jones <drjones@redhat.com>